server-django/requirements.txt
asgiref==3.2.10
bcrypt==3.2.0
cffi==1.14.3
Django==3.1.2
mysqlclient==2.0.1
pip-autoremove==0.9.1
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
